Extract CLI interface to own package (#18)
